companion/pyproject.toml
[project]
name = "student-task-hub-companion"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "browser-use>=0.4.0",
]

[project.scripts]
sth-companion = "sth_companion.main:main"

